Enable continuing from pickle
Python isn't my native language, so I may be wrong here, but it seemed to me, both from the code and in practice, that this ran in an infinite loop until you pressed 'n' or 'N', at which point the script exited. Actually continuing from the pickled contents was impossible?